refactor: 질문, 답변 도메인 패키지 이동
- question, answer -> qna 도메인으로 합침
