+++ src/main/java/shop/tryit/domain/qna/repository/QuestionRepository.java
@@ -1,0 +1,20 @@
+package shop.tryit.domain.qna.repository;
+
+import java.util.Optional;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
+import shop.tryit.domain.qna.dto.QuestionSearchCondition;
+import shop.tryit.domain.qna.dto.QuestionSearchDto;
+import shop.tryit.domain.qna.entity.Question;
+
+public interface QuestionRepository {
+
+    Long save(Question question);
+
+    Optional<Question> findById(Long id);
+
+    void delete(Question question);
+
+    Page<QuestionSearchDto> searchQuestion(QuestionSearchCondition condition, Pageable pageable);
+
+}
