--- src/main/java/shop/tryit/domain/question/dto/QuestionSearchCondition.java
@@ -1,18 +1,0 @@
-package shop.tryit.domain.question.dto;
-
-import lombok.Builder;
-import lombok.Getter;
-import lombok.ToString;
-
-@Getter
-@ToString
-public class QuestionSearchCondition {
-
-    private String condition;
-
-    @Builder
-    private QuestionSearchCondition(String condition, String email) {
-        this.condition = condition;
-    }
-
-}
