+++ src/main/java/shop/tryit/repository/qna/QuestionJpaRepository.java
@@ -1,0 +1,8 @@
+package shop.tryit.repository.qna;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import shop.tryit.domain.qna.entity.Question;
+
+public interface QuestionJpaRepository extends JpaRepository<Question, Long>, QuestionCustom {
+
+}
