--- src/main/java/shop/tryit/repository/question/QuestionCustom.java
@@ -1,12 +1,0 @@
-package shop.tryit.repository.question;
-
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.Pageable;
-import shop.tryit.domain.question.dto.QuestionSearchCondition;
-import shop.tryit.domain.question.dto.QuestionSearchDto;
-
-public interface QuestionCustom {
-
-    Page<QuestionSearchDto> searchQuestion(QuestionSearchCondition condition, Pageable pageable);
-
-}
